+++ backend/src/main/java/ua/edu/ukma/school_simplifier/repositories/ClassGroupRepository.java
@@ -3,5 +3,7 @@ package ua.edu.ukma.school_simplifier.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ua.edu.ukma.school_simplifier.domain.models.ClassGroup;
 
-public interface ClassGroupRepository extends JpaRepository<ClassGroup, Long> {
+import java.math.BigInteger;
+
+public interface ClassGroupRepository extends JpaRepository<ClassGroup, BigInteger> {
 }
