+++ backend/src/main/java/ua/edu/ukma/school_simplifier/repositories/LessonRepository.java
@@ -3,5 +3,7 @@ package ua.edu.ukma.school_simplifier.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ua.edu.ukma.school_simplifier.domain.models.Lesson;
 
-public interface LessonRepository extends JpaRepository<Lesson, Long> {
+import java.math.BigInteger;
+
+public interface LessonRepository extends JpaRepository<Lesson, BigInteger> {
 }
