+++ backend/src/main/java/ua/edu/ukma/school_simplifier/repositories/MarkBookRepository.java
@@ -3,5 +3,7 @@ package ua.edu.ukma.school_simplifier.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ua.edu.ukma.school_simplifier.domain.models.MarkBookRecord;
 
-public interface MarkBookRepository extends JpaRepository<MarkBookRecord, Long> {
+import java.math.BigInteger;
+
+public interface MarkBookRepository extends JpaRepository<MarkBookRecord, BigInteger> {
 }
