+++ backend/src/main/java/ua/edu/ukma/school_simplifier/repositories/PrincipalRepository.java
@@ -6,9 +6,10 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import ua.edu.ukma.school_simplifier.domain.security.Principal;
 
+import java.math.BigInteger;
 import java.util.Optional;
 
-public interface PrincipalRepository extends JpaRepository<Principal, Long> {
+public interface PrincipalRepository extends JpaRepository<Principal, BigInteger> {
 
     @Query(value = "FROM Principal u WHERE u.email = :email")
     Principal findPrincipalByEmail(@Param("email") String email);
@@ -16,13 +17,13 @@ public interface PrincipalRepository extends JpaRepository<Principal, Long> {
     @Modifying(clearAutomatically = true, flushAutomatically = true)
     @Query(value = "DELETE FROM principal_jwt WHERE principal_id = :target_principal_id",
             nativeQuery = true)
-    void deleteTokenForPrincipal(@Param("target_principal_id") Long principalId);
+    void deleteTokenForPrincipal(@Param("target_principal_id") BigInteger principalId);
 
     @Modifying(clearAutomatically = true, flushAutomatically = true)
     @Query(value = "INSERT INTO principal_jwt(principal_id, token) " +
                    "VALUES(:target_principal_id, :target_token)",
             nativeQuery = true)
-    void addTokenForPrincipal(@Param("target_principal_id") Long principalId,
+    void addTokenForPrincipal(@Param("target_principal_id") BigInteger principalId,
                                  @Param("target_token") String refreshToken);
 
     @Query(value = "SELECT token FROM principal_jwt " +
