+++ backend/src/main/java/ua/edu/ukma/school_simplifier/repositories/RoleRepository.java
@@ -5,7 +5,9 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import ua.edu.ukma.school_simplifier.domain.security.Role;
 
-public interface RoleRepository extends JpaRepository<Role, Long> {
+import java.math.BigInteger;
+
+public interface RoleRepository extends JpaRepository<Role, BigInteger> {
 
     @Query(value = "FROM Role WHERE roleName = :roleName")
     Role findRoleByName(@Param("roleName") String roleName);
