+++ backend/src/main/java/ua/edu/ukma/school_simplifier/repositories/ScheduleRepository.java
@@ -3,5 +3,7 @@ package ua.edu.ukma.school_simplifier.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ua.edu.ukma.school_simplifier.domain.models.ScheduleRecord;
 
-public interface ScheduleRepository extends JpaRepository<ScheduleRecord, Long> {
+import java.math.BigInteger;
+
+public interface ScheduleRepository extends JpaRepository<ScheduleRecord, BigInteger> {
 }
