+++ backend/src/main/java/ua/edu/ukma/school_simplifier/repositories/SchoolClassRepository.java
@@ -3,5 +3,7 @@ package ua.edu.ukma.school_simplifier.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ua.edu.ukma.school_simplifier.domain.models.SchoolClass;
 
-public interface SchoolClassRepository extends JpaRepository<SchoolClass, Long> {
+import java.math.BigInteger;
+
+public interface SchoolClassRepository extends JpaRepository<SchoolClass, BigInteger> {
 }
