+++ backend/src/main/java/ua/edu/ukma/school_simplifier/repositories/SubjectRepository.java
@@ -3,5 +3,7 @@ package ua.edu.ukma.school_simplifier.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ua.edu.ukma.school_simplifier.domain.models.Subject;
 
-public interface SubjectRepository extends JpaRepository<Subject, Long> {
+import java.math.BigInteger;
+
+public interface SubjectRepository extends JpaRepository<Subject, BigInteger> {
 }
