+++ backend/src/main/java/ua/edu/ukma/school_simplifier/repositories/TeacherRepository.java
@@ -3,5 +3,7 @@ package ua.edu.ukma.school_simplifier.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ua.edu.ukma.school_simplifier.domain.models.Teacher;
 
-public interface TeacherRepository extends JpaRepository<Teacher, Long> {
+import java.math.BigInteger;
+
+public interface TeacherRepository extends JpaRepository<Teacher, BigInteger> {
 }
